gera id aleatorio em todas apis

+++ TecnoMundo/TecnoMundo.CartAPI/Data/ValueObjects/CartDetailVO.cs
@@ -2,10 +2,10 @@
 {
     public class CartDetailVO
     {
-        public long Id { get; set; }
-        public long CartHeaderId { get; set; }
+        public int Id { get; set; }
+        public int CartHeaderId { get; set; }
         public CartHeaderVO? CartHeader { get; set; }
-        public long ProductId { get; set; }
+        public int ProductId { get; set; }
         public ProductVO Product { get; set; }
         public int Count { get; set; }
     }
