+++ TecnoMundo/TecnoMundo.CartAPI/Model/CartDetail.cs
@@ -6,10 +6,10 @@ namespace GeekShopping.CartAPI.Model
     [Table("cart_detail")]
     public class CartDetail : BaseEntity
     {
-        public long CartHeaderId { get; set; }
+        public int CartHeaderId { get; set; }
         [ForeignKey("CartHeaderId")]
         public virtual CartHeader CartHeader { get; set; }
-        public long ProductId { get; set; }
+        public int ProductId { get; set; }
         public virtual Product Product { get; set; }
         [Column("count")]
         public int Count { get; set; }
