+++ TecnoMundo/TecnoMundo.OrderAPI/Messages/CartDetailVO.cs
@@ -2,9 +2,9 @@
 {
     public class CartDetailVO
     {
-        public long Id { get; set; }
-        public long CartHeaderId { get; set; }
-        public long ProductId { get; set; }
+        public int Id { get; set; }
+        public int CartHeaderId { get; set; }
+        public int ProductId { get; set; }
         public virtual ProductVO Product { get; set; }
         public int Count { get; set; }
     }
