+++ TecnoMundo/TecnoMundo.OrderAPI/Model/OrderDetail.cs
@@ -6,11 +6,11 @@ namespace GeekShopping.OrderAPI.Model
     [Table("order_detail")]
     public class OrderDetail : BaseEntity
     {
-        public long OrderHeaderId { get; set; }
+        public int OrderHeaderId { get; set; }
         [ForeignKey("OrderHeaderId")]
         public virtual OrderHeader OrderHeader { get; set; }
         [Column("ProductId")]
-        public long ProductId { get; set; }
+        public int ProductId { get; set; }
         [Column("count")]
         public int Count { get; set; }
         [Column("product_name")]
